Move toolbox tag handling into tools
